updated todos

+++ src/Services/leaderService.ts
@@ -7,16 +7,21 @@ import { responseTypes } from "../Constants/responseTypes";
 export class LeaderService {
     constructor() { }
 
+    static hasMajority(): boolean {
+        // TODO: check if the leader has majority
+        return true
+    }
+
     static checkClientExists(data: any, sock: Socket): boolean {
         const { identity, serverid } = data
         // Check database
         if (ServiceLocator.foreignClientsDAO.isRegistered(identity)) {
-            writeJSONtoSocket(sock, { exists: true, type: responseTypes.IS_CLIENT, identity });
-        } else {
-            // TODO: check if the leader has majority
+            writeJSONtoSocket(sock, { acknowledged: true, exists: true, type: responseTypes.IS_CLIENT, identity });
+        // check if the leader has majority
+        } else if (this.hasMajority()) {
             ServiceLocator.serversDAO.incrementClock()
             ServiceLocator.foreignClientsDAO.addNewClient(serverid, identity)
-            writeJSONtoSocket(sock, { exists: false, type: responseTypes.IS_CLIENT, identity });
+            writeJSONtoSocket(sock, { acknowledged: true, exists: false, type: responseTypes.IS_CLIENT, identity });
             // Inform other servers
             LeaderService.broadcastServers({
                 type: responseTypes.BROADCAST_SERVER_UPDATE,
@@ -25,6 +30,8 @@ export class LeaderService {
                 clients: ServiceLocator.foreignClientsDAO.getClients(),
                 chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
             })
+        } else {
+            writeJSONtoSocket(sock, { acknowledged: false, exists: false, type: responseTypes.IS_CLIENT, identity });
         }
         return true
     }
@@ -33,12 +40,12 @@ export class LeaderService {
         const { roomid, serverid } = data
         // Check database
         if (ServiceLocator.foreignChatroomsDAO.isRegistered(roomid)) {
-            writeJSONtoSocket(sock, { exists: true, type: responseTypes.IS_CHATROOM, roomid });
-        } else {
-            // TODO: check if the leader has majority
+            writeJSONtoSocket(sock, { acknowledged: true, exists: true, type: responseTypes.IS_CHATROOM, roomid });
+        // check if the leader has majority
+        } else if (this.hasMajority()) {
             ServiceLocator.serversDAO.incrementClock()
             ServiceLocator.foreignChatroomsDAO.addNewChatroom(serverid, roomid)
-            writeJSONtoSocket(sock, { exists: false, type: responseTypes.IS_CHATROOM, roomid });
+            writeJSONtoSocket(sock, { acknowledged: true, exists: false, type: responseTypes.IS_CHATROOM, roomid });
             // Inform other servers
             LeaderService.broadcastServers({
                 type: responseTypes.BROADCAST_SERVER_UPDATE,
@@ -47,6 +54,8 @@ export class LeaderService {
                 clients: ServiceLocator.foreignClientsDAO.getClients(),
                 chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
             })
+        } else {
+            writeJSONtoSocket(sock, { acknowledged: false, exists: false, type: responseTypes.IS_CHATROOM, roomid });
         }
         return true
     }
@@ -75,35 +84,43 @@ export class LeaderService {
 
     static acknowledgeChatroomDeletion(data: any, sock: Socket): boolean {
         const { roomid, serverid } = data
-        // TODO: check if the leader has majority
-        ServiceLocator.serversDAO.incrementClock()
-        ServiceLocator.foreignChatroomsDAO.removeChatroom(serverid, roomid);
-        writeJSONtoSocket(sock, { acknowledged: true, type: responseTypes.INFORM_ROOMDELETION, roomid });
-        // inform other servers
-        LeaderService.broadcastServers({
-            type: responseTypes.BROADCAST_SERVER_UPDATE,
-            leaderid: ServiceLocator.serversDAO.getLeaderId(),
-            clock: ServiceLocator.serversDAO.getClock(),
-            clients: ServiceLocator.foreignClientsDAO.getClients(),
-            chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
-        })
+        // check if the leader has majority
+        if (this.hasMajority()) {
+            ServiceLocator.serversDAO.incrementClock()
+            ServiceLocator.foreignChatroomsDAO.removeChatroom(serverid, roomid);
+            writeJSONtoSocket(sock, { acknowledged: true, type: responseTypes.INFORM_ROOMDELETION, roomid });
+            // inform other servers
+            LeaderService.broadcastServers({
+                type: responseTypes.BROADCAST_SERVER_UPDATE,
+                leaderid: ServiceLocator.serversDAO.getLeaderId(),
+                clock: ServiceLocator.serversDAO.getClock(),
+                clients: ServiceLocator.foreignClientsDAO.getClients(),
+                chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
+            })
+        } else {
+            writeJSONtoSocket(sock, { acknowledged: false, type: responseTypes.INFORM_ROOMDELETION, roomid });
+        }
         return true
     }
 
     static acknowledgeClientDeletion(data: any, sock: Socket): boolean {
         const { identity, serverid } = data
-        // TODO: check if the leader has majority
-        ServiceLocator.serversDAO.incrementClock()
-        ServiceLocator.foreignClientsDAO.removeClient(serverid, identity);
-        writeJSONtoSocket(sock, { acknowledged: true, type: responseTypes.INFORM_CLIENTDELETION, identity });
-        // inform other servers
-        LeaderService.broadcastServers({
-            type: responseTypes.BROADCAST_SERVER_UPDATE,
-            leaderid: ServiceLocator.serversDAO.getLeaderId(),
-            clock: ServiceLocator.serversDAO.getClock(),
-            clients: ServiceLocator.foreignClientsDAO.getClients(),
-            chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
-        })
+        // check if the leader has majority
+        if (this.hasMajority()) {
+            ServiceLocator.serversDAO.incrementClock()
+            ServiceLocator.foreignClientsDAO.removeClient(serverid, identity);
+            writeJSONtoSocket(sock, { acknowledged: true, type: responseTypes.INFORM_CLIENTDELETION, identity });
+            // inform other servers
+            LeaderService.broadcastServers({
+                type: responseTypes.BROADCAST_SERVER_UPDATE,
+                leaderid: ServiceLocator.serversDAO.getLeaderId(),
+                clock: ServiceLocator.serversDAO.getClock(),
+                clients: ServiceLocator.foreignClientsDAO.getClients(),
+                chatrooms: ServiceLocator.foreignChatroomsDAO.getChatrooms()
+            })
+        } else {
+            writeJSONtoSocket(sock, { acknowledged: false, type: responseTypes.INFORM_CLIENTDELETION, identity });
+        }
         return true
     }
